// lib/config/interfaces/environment-config.ts
export interface FsxOntapConfig {
  enabled: boolean;
  storageCapacity: number;
  throughputCapacity: number;
  deploymentType: string;
  automaticBackupRetentionDays: number;
  volumes?: any;
  activeDirectory?: {
    enabled: boolean;
    domainName?: string;
    dnsIps?: string[];
  };
}